requirements.txt
annotated-types==0.7.0
anyio==4.3.0
bcrypt==4.0.1
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.7.1
pydantic_core==2.18.2
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.12.0
uvicorn==0.29.0
